+++ packages/yourloops/lib/config/config.ts
@@ -65,7 +65,8 @@ const defaultConfig: AppConfig = {
   YLPZ_RA_LAD_IT: 'YLPZ-RA-LAD-001-it-Rev6',
   YLPZ_RA_LAD_ES: 'YLPZ-RA-LAD-001-es-Rev6',
   YLPZ_RA_LAD_DE: 'YLPZ-RA-LAD-001-de-Rev6',
-  IDLE_TIMEOUT_MS: 1800000
+  IDLE_TIMEOUT_MS: 1800000,
+  EATING_SHORTLY_ENABLED: false
 }
 const appConfig = _.assign({}, defaultConfig)
 if (_.has(window, 'config') && _.isObjectLike(_.get(window, 'config', null))) {